double

+++ Models/Balances.cs
@@ -8,6 +8,6 @@ namespace PersonalFinance.Models
         public string Usr_OID { get; set; }
         public int ID { get; set; }
         public DateTime BalDateTime { get; set; }
-        public float ActBalance { get; set; }
+        public double ActBalance { get; set; }
     }
 }
